python/requirements_of_rag.txt
protobuf==3.20.3
pydantic==1.10.13
chromadb==0.4.18
sentence-transformers==2.2.2
huggingface-hub==0.19.4
groq==0.9.0
python-dotenv==1.0.0
tqdm==4.66.0
numpy==1.24.3
requests==2.31.0
setuptools==75.5.0
wheel==0.44.0